Fix preserving of id's in commands and events
Sending a command which is in reaction to an event, should preserver
it's `process_id`, which was created when first command, a blank
command that was created because of the external action. Similar with
`predecessor_id` which should be the ID only of events, not of commands
since command log is not stored in eventsourcing.