Publish events from spawned process
When events are published, after they were persisted, they are sent from
a newly spawned process, not linked to anything. To guarantee it will
not be killed by sudden death of a client which would resulted in a
inconsistency between readstores or process managers.

Additionally release AR when command returned error.